+++ modules/spring-web/src/main/kotlin/com/esprito/spring/web/view/EndpointsDocumentationDetailPanel.kt
@@ -1,0 +1,9 @@
+package com.esprito.spring.web.view
+
+import com.intellij.openapi.Disposable
+import com.intellij.openapi.project.Project
+
+class EndpointsDocumentationDetailPanel(project: Project) : Disposable {
+    override fun dispose() {
+    }
+}